+++ apps/melding-form/src/app/(general)/aanvullende-vragen/[classificationId]/[panelId]/actions.ts
@@ -3,8 +3,9 @@
 import { cookies } from 'next/headers'
 import { redirect } from 'next/navigation'
 
-import { postMeldingByMeldingIdQuestionByQuestionId, putMeldingByMeldingIdAnswerQuestions } from '@meldingen/api-client'
+import { putMeldingByMeldingIdAnswerQuestions } from '@meldingen/api-client'
 
+import { buildAnswerPromises } from './_utils/buildAnswerPromises'
 import { mergeCheckboxAnswers } from './_utils/mergeCheckboxAnswers'
 import { handleApiError } from 'apps/melding-form/src/handleApiError'
 
@@ -34,37 +35,6 @@ const mapValidationErrors = (keys: string[]) =>
     key,
     message: 'Vraag is verplicht en moet worden beantwoord.',
   }))
-
-export const buildAnswerPromises = (
-  entries: [string, string | File][],
-  questionKeysAndIds: { key: string; id: number }[],
-  meldingId: string,
-  token: string,
-) =>
-  entries.map(([key, value]) => {
-    if (value instanceof File) return undefined
-
-    // Filter out empty answers
-    if (value.length === 0) return undefined
-
-    const questionId = questionKeysAndIds.find((component) => component.key === key)?.id
-
-    if (!questionId) return undefined
-
-    return postMeldingByMeldingIdQuestionByQuestionId({
-      body: { text: value },
-      path: {
-        melding_id: parseInt(meldingId, 10),
-        question_id: questionId,
-      },
-      query: { token },
-    }).then((results) => {
-      return {
-        key: key,
-        value: results,
-      }
-    })
-  })
 
 export const postForm = async (
   { isLastPanel, lastPanelPath, nextPanelPath, questionKeysAndIds, requiredQuestionKeys }: ArgsType,
